requirements.txt
# This template is a low-dependency template.
# By default there is no requirements added here.
# Add the requirements you need to this file.
# or run `make init` to create this file automatically based on the template.
# You can also run `make switch-to-poetry` to use the poetry package manager.

podgen
gistyc
python-slugify
python-dotenv
lxml_html_clean
dropbox
scrapy
omegaconf
